rename the typography package to text cleanup all usages fix docs
