+++ packages/@react-spectrum/button/docs/ActionButton.mdx
@@ -37,7 +37,7 @@ category: Buttons
 ## Example
 
 ```tsx example
-import {Text} from '@react-spectrum/typography';
+import {Text} from '@react-spectrum/text';
 
 <ActionButton>
   <Edit />
