+++ packages/@react-spectrum/button/docs/Button.mdx
@@ -47,7 +47,7 @@ Buttons can have a label, and icon, or both. An icon is provided by passing an i
 A visible label can be provided by passing children.
 
 ```tsx example
-import {Text} from '@react-spectrum/typography';
+import {Text} from '@react-spectrum/text';
 
 <Button variant="primary">Label only</Button>
 <Button variant="primary">
