+++ packages/@react-spectrum/button/src/ActionButton.tsx
@@ -17,7 +17,7 @@ import {FocusRing} from '@react-aria/focus';
 import React from 'react';
 import {SpectrumActionButtonProps} from '@react-types/button';
 import styles from '@adobe/spectrum-css-temp/components/button/vars.css';
-import {Text} from '@react-spectrum/typography';
+import {Text} from '@react-spectrum/text';
 import {useButton} from '@react-aria/button';
 import {useProviderProps} from '@react-spectrum/provider';
