+++ packages/@react-spectrum/button/stories/ActionButton.stories.tsx
@@ -15,7 +15,7 @@ import {ActionButton} from '../';
 import Add from '@spectrum-icons/workflow/Add';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
-import {Text} from '@react-spectrum/typography';
+import {Text} from '@react-spectrum/text';
 
 storiesOf('Button/ActionButton', module)
   .addParameters({providerSwitcher: {status: 'positive'}})
