+++ packages/@react-spectrum/buttongroup/package.json
@@ -32,7 +32,7 @@
     "@react-aria/utils": "^3.0.0-rc.2",
     "@react-spectrum/utils": "^3.0.0-rc.2",
     "@react-spectrum/button": "^3.0.0-rc.2",
-    "@react-spectrum/typography": "^3.0.0-alpha.1",
+    "@react-spectrum/text": "^3.0.0-alpha.1",
     "@react-types/buttongroup": "^3.0.0-alpha.1",
     "@react-types/shared": "^3.0.0-rc.2"
   },
