+++ packages/@react-spectrum/dialog/docs/Dialog.mdx
@@ -50,7 +50,7 @@ import {ButtonGroup} from '@react-spectrum/buttongroup';
 import {Content, Header} from '@react-spectrum/view';
 import {Dialog, DialogTrigger} from '@react-spectrum/dialog';
 import {Divider} from '@react-spectrum/divider';
-import {Heading, Text} from '@react-spectrum/typography';
+import {Heading, Text} from '@react-spectrum/text';
 
 <DialogTrigger>
   <ActionButton>Check connectivity</ActionButton>
