+++ packages/@react-spectrum/dialog/docs/DialogTrigger.mdx
@@ -22,7 +22,7 @@ import {Content, Footer, Header} from '@react-spectrum/view';
 import {Dialog, DialogTrigger} from '@react-spectrum/dialog';
 import {Divider} from '@react-spectrum/divider';
 import {Form} from '@react-spectrum/form';
-import {Heading, Text} from '@react-spectrum/typography';
+import {Heading, Text} from '@react-spectrum/text';
 import {TextField} from '@react-spectrum/textfield';
 import {Flex} from '@react-spectrum/layout';
 ```
