+++ packages/@react-spectrum/dialog/package.json
@@ -40,7 +40,7 @@
     "@react-spectrum/divider": "^3.0.0-rc.2",
     "@react-spectrum/layout": "^3.0.0-alpha.1",
     "@react-spectrum/overlays": "^3.0.0-alpha.1",
-    "@react-spectrum/typography": "^3.0.0-alpha.1",
+    "@react-spectrum/text": "^3.0.0-alpha.1",
     "@react-spectrum/utils": "^3.0.0-rc.2",
     "@react-spectrum/view": "^3.0.0-alpha.1",
     "@react-stately/dialog": "^3.0.0-alpha.1",
