+++ packages/@react-spectrum/divider/docs/Divider.mdx
@@ -17,7 +17,7 @@ import packageData from '../package.json';
 ```jsx import
 import {Divider} from '@react-spectrum/divider';
 import {Flex} from '@react-spectrum/layout';
-import {Text} from '@react-spectrum/typography';
+import {Text} from '@react-spectrum/text';
 ```
 
 ---
