+++ packages/@react-spectrum/illustratedmessage/docs/IllustratedMessage.mdx
@@ -16,7 +16,7 @@ import packageData from '../package.json';
 
 ```jsx import
 import {Content} from '@react-spectrum/view';
-import {Heading} from '@react-spectrum/typography';
+import {Heading} from '@react-spectrum/text';
 import {IllustratedMessage} from '@react-spectrum/illustratedmessage';
 ```
