+++ packages/@react-spectrum/illustratedmessage/test/IllustratedMessage.test.js
@@ -11,7 +11,7 @@
  */
 
 import {Content} from '@react-spectrum/view';
-import {Heading} from '@react-spectrum/typography';
+import {Heading} from '@react-spectrum/text';
 import {IllustratedMessage} from '../';
 import React from 'react';
 import {render} from '@testing-library/react';
