+++ packages/@react-spectrum/listbox/package.json
@@ -38,7 +38,7 @@
     "@react-spectrum/layout": "^3.0.0-alpha.1",
     "@react-spectrum/progress": "^3.0.0-rc.2",
     "@react-spectrum/provider": "^3.0.0-rc.2",
-    "@react-spectrum/typography": "^3.0.0-alpha.1",
+    "@react-spectrum/text": "^3.0.0-alpha.1",
     "@react-spectrum/utils": "^3.0.0-rc.2",
     "@react-stately/collections": "^3.0.0-alpha.1",
     "@react-stately/list": "^3.0.0-alpha.1",
