+++ packages/@react-spectrum/listbox/stories/ListBox.stories.tsx
@@ -23,7 +23,7 @@ import {Label} from '@react-spectrum/label';
 import Paste from '@spectrum-icons/workflow/Paste';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
-import {Text} from '@react-spectrum/typography';
+import {Text} from '@react-spectrum/text';
 import {useAsyncList} from '@react-stately/data';
 
 let iconMap = {
