+++ packages/@react-spectrum/menu/docs/MenuTrigger.mdx
@@ -195,7 +195,7 @@ A Menu Item's content may be any renderable node, not just strings.
 [View guidelines](https://spectrum.adobe.com/page/popovers/#Content-types)
 
 ```tsx example
-import {Keyboard, Text} from '@react-spectrum/typography';
+import {Keyboard, Text} from '@react-spectrum/text';
 
 <MenuTrigger>
   <ActionButton>
