+++ packages/@react-spectrum/menu/package.json
@@ -41,7 +41,7 @@
     "@react-spectrum/divider": "^3.0.0-rc.2",
     "@react-spectrum/layout": "^3.0.0-alpha.1",
     "@react-spectrum/overlays": "^3.0.0-alpha.1",
-    "@react-spectrum/typography": "^3.0.0-alpha.1",
+    "@react-spectrum/text": "^3.0.0-alpha.1",
     "@react-spectrum/utils": "^3.0.0-rc.2",
     "@react-stately/collections": "^3.0.0-alpha.1",
     "@react-stately/menu": "^3.0.0-alpha.1",
