+++ packages/@react-spectrum/picker/docs/Picker.mdx
@@ -22,7 +22,7 @@ import BulkEditUsers from '@spectrum-icons/workflow/BulkEditUsers';
 import Draw from '@spectrum-icons/workflow/Draw';
 import ColorPalette from '@spectrum-icons/workflow/ColorPalette';
 import AlignLeft from '@spectrum-icons/workflow/AlignLeft';
-import {Text} from '@react-spectrum/typography';
+import {Text} from '@react-spectrum/text';
 ```
 
 ---
