+++ packages/@react-spectrum/picker/package.json
@@ -41,7 +41,7 @@
     "@react-spectrum/overlays": "^3.0.0-alpha.1",
     "@react-spectrum/progress": "^3.0.0-rc.2",
     "@react-spectrum/provider": "^3.0.0-rc.2",
-    "@react-spectrum/typography": "^3.0.0-alpha.1",
+    "@react-spectrum/text": "^3.0.0-alpha.1",
     "@react-spectrum/utils": "^3.0.0-rc.2",
     "@react-stately/select": "^3.0.0-alpha.1",
     "@react-types/select": "^3.0.0-alpha.1",
