+++ packages/@react-spectrum/text/docs/Keyboard.mdx
@@ -10,43 +10,44 @@ governing permissions and limitations under the License. -->
 import {Layout} from '@react-spectrum/docs';
 export default Layout;
 
-import docs from 'docs:@react-spectrum/typography';
+import docs from 'docs:@react-spectrum/text';
 import {HeaderInfo, PropTable} from '@react-spectrum/docs';
 import packageData from '../package.json';
 
 ```jsx import
-import {Text} from '@react-spectrum/typography';
+import {Heading, Keyboard, Text} from '@react-spectrum/text';
 ```
 
 ---
 category: Content
 ---
 
-# Text
+# Keyboard
 
-<p>{docs.exports.Text.description}</p>
+<p>{docs.exports.Keyboard.description}</p>
 
 <HeaderInfo
   packageData={packageData}
-  componentNames={['Text']} />
+  componentNames={['Keyboard']} />
 
 ## Example
 
 ```tsx example
-<Text>Paste</Text>
+<Keyboard>⌘V</Keyboard>
 ```
 
 ## Content
 
-Text is related to typography components and is a child to other
-components. Text is used to wrap text strings and elements within other
-components. Based on the component it is used within, it gets Spectrum styling
+Keyboard is used to format text indicating that is a keyboard
+command. Based on the component it is used within, it gets Spectrum styling
 and may have a preassigned slot for positioning. By virtue of its name, the
-Text component provides semantic meaning in the context they are placed.
-Examples of its usage in [Dialog](Dialog.html#example) and
-[Button](Button.html#content). It is important to note that the Text component
-accepts any renderable node, not just strings.
+Keyboard component provides semantic meaning in the context they are placed.
+An example of its usage in [Menu](MenuTrigger.html#complex-menu-items).
+It is important to note that the Keyboard component accepts any renderable
+node, not just strings.
+
+See [the MDN docs](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/kbd) for more information.
 
 ## Props
 
-<PropTable component={docs.exports.Text} links={docs.links} />
+<PropTable component={docs.exports.Keyboard} links={docs.links} />
