+++ packages/@react-spectrum/well/docs/Well.mdx
@@ -15,7 +15,7 @@ import {HeaderInfo, PropTable} from '@react-spectrum/docs';
 import packageData from '../package.json';
 
 ```jsx import
-import {Heading} from '@react-spectrum/typography';
+import {Heading} from '@react-spectrum/text';
 import {Well} from '@react-spectrum/well';
 ```
