+++ packages/@react-types/text/README.md
@@ -1,3 +1,3 @@
-# @react-types/typography
+# @react-types/text
 
 This package is part of [react-spectrum](https://github.com/adobe-private/react-spectrum-v3). See the repo for more details.
