+++ packages/@react-types/text/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@react-types/typography",
+  "name": "@react-types/text",
   "version": "3.0.0-alpha.1",
   "description": "Spectrum UI components in React",
   "license": "Apache-2.0",
