+++ packages/@react-types/text/src/heading.d.ts
